Oops, I didn't actually fix that warning.

+++ src/mlpack/methods/emst/dtb_impl.hpp
@@ -268,7 +268,7 @@ void DualTreeBoruvka<MetricType, TreeType>::CleanupHelper(TreeType* tree)
 
   // Check components of points.
   for (size_t i = 0; i < tree->NumPoints(); ++i)
-    if (connections.Find(tree->Point(i)) != int(component))
+    if (connections.Find(tree->Point(i)) != size_t(component))
       return;
 
   // If we made it this far, all components are the same.
